+++ packages/platform-android/src/adb-failure.ts
@@ -112,7 +112,7 @@ const ANDROID_ADB_FAILURE_MATCHERS = [
   ],
 ] as const satisfies readonly AndroidAdbFailureMatcher[];
 
-export const ANDROID_ADB_TIMEOUT_FAILURE: AndroidAdbFailureClassification = Object.freeze({
+const ANDROID_ADB_TIMEOUT_FAILURE: AndroidAdbFailureClassification = Object.freeze({
   reason: 'timeout',
   hint: 'adb timed out — the adb server may be wedged. Run adb kill-server && adb start-server, check adb devices, then retry.',
 });
@@ -133,6 +133,81 @@ export function classifyAndroidAdbFailure(
 
 import { AppError } from '@agent-device/kernel/errors';
 import type { HostCommandResult } from '@agent-device/contracts/platform-runtime-host';
+import type { AndroidAdbExecutorResult } from './adb-transport.ts';
+import { requireAndroidAdbHost } from './adb-host.ts';
+
+/**
+ * Enriches a failed adb command error in place with the classified hint,
+ * `retriable` flag, and machine-readable `adbFailure` family, so every adb call
+ * site surfaces guidance without per-site classification. Exec-layer timeouts
+ * classify as `timeout` even though they leave no stderr. No-op for errors that
+ * are not adb command failures or that carry no recognized failure signal; an
+ * existing hint or retriable verdict is never overwritten.
+ */
+export function attachAdbFailureHint<T>(error: T): T {
+  if (!(error instanceof AppError) || error.code !== 'COMMAND_FAILED') return error;
+  const classification = classifyAdbCommandError(error);
+  if (!classification) return error;
+  error.details = {
+    ...error.details,
+    adbFailure: classification.reason,
+    ...(typeof error.details?.hint === 'string' ? {} : { hint: classification.hint }),
+    ...(classification.retriable !== undefined && error.details?.retriable === undefined
+      ? { retriable: classification.retriable }
+      : {}),
+  };
+  return error;
+}
+
+// Timeout wins over text matchers: the exec layer deliberately builds timeout
+// errors around "timed out after Nms" because partial output from the killed
+// process is untrustworthy — classifying that partial stderr (e.g. flagging a
+// half-written transport line retriable) would point away from the real fix.
+function classifyAdbCommandError(error: AppError): AndroidAdbFailureClassification | undefined {
+  if (typeof error.details?.timeoutMs === 'number') return ANDROID_ADB_TIMEOUT_FAILURE;
+  const stderr = typeof error.details?.stderr === 'string' ? error.details.stderr : '';
+  const stdout = typeof error.details?.stdout === 'string' ? error.details.stdout : '';
+  return classifyAndroidAdbFailure(stderr, stdout);
+}
+
+/**
+ * Builds the COMMAND_FAILED AppError for a failed `allowFailure` adb result and
+ * runs it through the failure classifier. This is the shared construction point
+ * for call sites that tolerate a failure to inspect it and then throw — those
+ * errors never cross the executor throw path, so they classify here instead.
+ * Site-provided `details` win on key collisions, and a site `hint` is preserved
+ * over the classified one.
+ *
+ * Nonzero exits build their details via the host's execFailureDetails, whose
+ * processExitError flag makes normalizeError append the first stderr line to
+ * the curated message — the classified hint and the stderr-excerpt enrichment
+ * compose instead of competing. Semantic failures thrown at exit 0 (e.g. an
+ * `am start` error printed on a successful exit) stay unflagged so a stray
+ * stderr line never decorates a message the process exit does not back up.
+ */
+export function androidAdbResultError(
+  message: string,
+  result: Pick<AndroidAdbExecutorResult, 'exitCode' | 'stdout' | 'stderr'>,
+  details?: Record<string, unknown>,
+): AppError {
+  const failureDetails =
+    result.exitCode === 0
+      ? { stdout: result.stdout, stderr: result.stderr, exitCode: result.exitCode, ...details }
+      : requireAndroidAdbHost().execFailureDetails(result, details);
+  return attachAdbFailureHint(new AppError('COMMAND_FAILED', message, failureDetails));
+}
+
+export function withAdbFailureHints<Args extends unknown[], Result>(
+  call: (...args: Args) => Promise<Result>,
+): (...args: Args) => Promise<Result> {
+  return async (...args) => {
+    try {
+      return await call(...args);
+    } catch (error) {
+      throw attachAdbFailureHint(error);
+    }
+  };
+}
 
 export function androidDiscoveryCommandError(
   message: string,
